requirements.txt
pytest>=8.3.3
selene>=2.0.0rc9
allure-pytest>=2.13.5
requests>=2.32.3